+++ packages/files-ui/src/Components/Modules/FileBrowsers/views/FileSystemItem/SharedFolderRow.tsx
@@ -22,6 +22,7 @@ import { Form, FormikProvider, useFormik } from "formik"
 import { object, string } from "yup"
 import clsx from "clsx"
 import { BucketKeyPermission } from "../../../../../Contexts/FilesContext"
+import UserBubble from "../../../../Elements/UserBubble"
 
 const useStyles = makeStyles(({ breakpoints, constants, palette }: CSFTheme) => {
 
@@ -218,7 +219,9 @@ const SharedFolderRow = ({ bucket, onFolderClick, menuItems, isEditing, setIsEdi
       </TableCell>
       {desktop &&
         <TableCell align="left">
-          {isOwner && t`me`}
+          {isOwner
+            ? t`me`
+            : <UserBubble tooltip={bucket.owners[0].uuid || ""} />}
         </TableCell>
       }
       <TableCell
